+++ test/goldens/golden_homescreen_test.dart
@@ -1,19 +1,28 @@
+import 'dart:io';
 import 'dart:ui';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:gato_id_generator/src/app.dart';
-import 'package:integration_test/integration_test.dart';
+import 'package:gato_id_generator/src/core/_core.dart';
 
 import '../robot.dart';
 
 void main() {
+  setUp(() async {
+    try {
+      await Directory('temp').delete(recursive: true);
+      getIt.reset(dispose: true);
+    } catch (e) {
+      // print("ERROR");
+      // print(e);
+    }
+  });
+
   final sizeVariant = ValueVariant<Size>({
     const Size(300, 600),
     const Size(600, 800),
     const Size(1000, 1000),
   });
-
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets(
     'Golden - HomeScreen',
@@ -23,8 +32,7 @@ void main() {
       await r.golden.setSurfaceSize(currentSize);
       await r.golden.loadTextFont();
       await r.golden.loadMaterialIconFont();
-      await r.golden.loadBoxIconFont();
-      await r.pumpApp();
+      await tester.runAsync(() async => await r.pumpApp());
       await r.golden.precacheImages();
       await expectLater(
         find.byType(App),
@@ -36,7 +44,6 @@ void main() {
     variant: sizeVariant,
     tags: ['golden'],
     // TODO Skip this test until we can run it successfully on CI without this error:
-    // Golden "products_list_300x600.png": Pixel test failed, 2.33% diff detected.
     skip: false,
   );
 }