+++ test/robot.dart
@@ -1,21 +1,25 @@
 // ignore_for_file: depend_on_referenced_packages
 
 import 'dart:io';
+import 'dart:math';
 
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:gato_id_generator/src/core/local_db/hive_registrar.g.dart';
+import 'package:gato_id_generator/src/presentation/_common_widgets/custom_appbar.dart';
+import 'package:gato_id_generator/src/presentation/home/view/home_screen.dart';
 import 'package:hive_ce/hive.dart';
+import 'package:path_provider/path_provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:gato_id_generator/src/core/constants/_constants.dart';
 import 'package:gato_id_generator/src/core/_core.dart';
 
 import 'goldens/golden_robot.dart';
-import 'src/presentation/auth/auth_robot.dart';
-import 'src/presentation/generate/generate_robot.dart';
-import 'src/presentation/home/home_robot.dart';
+import 'robots/auth_robot.dart';
+import 'robots/generate_robot.dart';
+import 'robots/home_robot.dart';
 
 @visibleForTesting
 class Robot {
@@ -31,7 +35,7 @@ class Robot {
   final AuthRobot auth;
   final HomeRobot home;
 
-  Future<void> _necessaryIntializations() async {
+  Future<void> _necessaryIntializations(bool isIntegration) async {
     try {
       await EasyLocalization.ensureInitialized();
       Default.init();
@@ -39,31 +43,36 @@ class Robot {
       await AppInfo.init(const PackageInfoWrapper());
 
       // Hive-specific
-      Hive.init(Directory("temp").path);
+      Hive.init((isIntegration)
+          ? (await getTemporaryDirectory()).path
+          : Directory("temp/${Random().nextInt(10000)}/").path);
+
       Hive.registerAdapters();
       await Hive.initBoxes();
     } catch (e) {
       if (kDebugMode) {
         print("ERROR");
-        // print(e);
+        print(e);
       }
     }
   }
 
-  Future<void> pumpApp() async {
+  Future<void> pumpApp({bool isIntegration = false}) async {
     TestWidgetsFlutterBinding.ensureInitialized();
     WidgetsFlutterBinding.ensureInitialized();
 
     // Simulation requirements.
+    randomGatoCardPreview = false;
     canListenToNetworkStatusChange = false;
 
     // Easy_localization test requirement.
     SharedPreferences.setMockInitialValues({});
 
     // ! Must be before appStartup.
-    await _necessaryIntializations();
+    await _necessaryIntializations(isIntegration);
+
+    // Try to reset first.
     try {
-      // Try to reset first.
       await getIt.reset(dispose: true);
     } catch (e) {
       // Nothing
@@ -75,12 +84,30 @@ class Robot {
       await appStartup.initializeApp(minimumTest: true);
       await appStartup.runWithLocal();
     } catch (e) {
-      // Do nothing, just to prevent multiple getIt register error.
+      print("ERRRRRRROOOOOOOOOOORRRRRRRRRR");
+      print(e);
     }
 
     // * Entry point of the app
     final root = await appStartup.createRootWidget();
     await tester.pumpWidget(root);
+    await tester.pumpAndSettle();
+  }
+
+  Finder expectSnackBar() {
+    final finder = find.byType(SnackBar);
+    expect(finder, findsOneWidget);
+    return finder;
+  }
+
+  Finder expectBackButton() {
+    final finder = find.byType(BackIconButton);
+    expect(finder, findsOneWidget);
+    return finder;
+  }
+
+  Future<void> tapBackButton() async {
+    await tester.tap(expectBackButton());
     await tester.pumpAndSettle();
   }
 }