+++ test/src/presentation/generate/view/generate_screen_test.dart
@@ -29,4 +29,25 @@ void main() {
       await r.generate.expectInitialLayoutIsCorrect();
     });
   });
+
+  // ? Oddly the auth state is remembered after you run the first test above.
+  // TODO fix this?
+  testWidgets('''
+    Given in GenerateScreen, 
+    When not generating anything yet,
+    Then save button should not do anything when tapped.
+  ''', (tester) async {
+    await tester.runAsync(() async {
+      final r = Robot(tester);
+      await r.pumpApp();
+      await r.home.tapGenerateGatoButton();
+      // await r.auth.registerAndSignIn();
+      // await r.generate.expectInitialLayoutIsCorrect();
+
+      await r.generate.tapSaveGeneratedButton();
+      r.generate.expectHUDLoading(false);
+    });
+  });
+
+// TODO complete this.
 }