+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Widget tests.
 - Integration tests.
+- Golden tests.
 
 ## [1.2.1] - 2025-04-12